pyproject.toml
[project]
name = "image_matting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "albumentations>=2.0.8",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-box>=7.3.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]
